soc: dpio: invalidate EQCR entry before writing to it
By invalidating the cache line containing the EQCR entry that we are
about to write to we make sure that no dirty data is kept. This should
not have happened since we flush the entry after we write the verb.
This is probably just ignoring the root cause and not the final
solution.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>