can: flexcan: Fix can Tx interrupt not generating
CAN interrupt mask register decides, for which MBs the interrupt would
be generated.
we use last MB for TX.
when we operate in CAN 2.0 mode (non fd mode), total MBs are 64 MBs.
i.e. MB0 - MB63. MB63 is used for TX interrupts in this case.
i.e. we need to set BIT 63, out of total 64 bits of
reg_imask2 << 32 | reg_imask1

Because using BIT(x) for setting TX interrupt mask, which only sets bits
in 32 bit mode, this caused interrupt mask for TX MB to not set.
Which results in TX Interrupts being not generated.

Fix this by using BIT_ULL which can set the bit 63 for tx MB.

Fixes: ab6467481657 ("can: flexcan: modify the interrupt register handling")

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>