drivers: net: can: flexcan: Fix incorrect Tx messages
The commit 523a754dba64 changed the can frame to canfd frames for all Tx
and Rx operations.
It introduced a bug that only can_dlc amount of bytes were transferred
instead of whole frame. in CAN 2.0 the can_dlc and can_len are same, which
is why this bug went unnoticed in CAN 2.0 mode. This bug would only appear
when a CAN FD message is being transmitted.

Fix the bug by correctly using the can_len while transmitting the frame.

Fixes: 523a754dba64 ("net: can: flexcan: use CAN FD frames for Tx/Rx")

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>