enetc: Fix enetc VF module dependency on FSL_ENETC_MDIO
[config part]

This is another workaround for the MDIO h/w issue. The MDIO
issue requires a global lock, which in turn creates linkage
errors for the ENETC modules when built as standalone dynamic
modules.
This patch fixes the mentioned linkage error by moving the
lock to the mdio module and making the vf driver too
to depend explicitly on the mdio module.

Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Integrated-by: Zhao Qiang <qiang.zhao@nxp.com>