crypto: caam - add power management support for Layerscape platforms
In Power Management in order for the device to go into sleep(LPM20)
it needs to quiesce all internal IP modules. We quiesce the SEC module
by masking the interrupts in all job ring during suspend. For sleep there
is no need for flushing the job ring so in resume we only enable the
interrupts back.

Signed-off-by: Andrei Botila <andrei.botila@nxp.com>